+++ Codemine/Extensions/String+EmailValidation.swift
@@ -1,0 +1,25 @@
+//
+//  String+EmailValidation.swift
+//  Codemine
+//
+//  Created by Marius Constantinescu on 18/02/16.
+//  Copyright © 2016 Nodes. All rights reserved.
+//
+
+import Foundation
+
+
+public extension String {
+ 
+    /**
+     Checks if the current string is a valid email address.
+     
+     - returns: true is the current string is a valid email address, false otherwise
+     */
+    public func isValidEmailAddress() -> Bool {
+        
+        let emailRegex = "\\A[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\z"
+        let emailTest = NSPredicate(format:"SELF MATCHES %@", emailRegex)
+        return emailTest.evaluateWithObject(self)
+    }
+}
